refactor file structure for new LBP
