+++ src/newLaunch/lbp/stage3.html
@@ -1,6 +1,6 @@
 <template>
-  <require from="../resources/elements/questionMark/questionMark.html"></require>
-  <require from="./tokenLabel.html"></require>
+  <require from="resources/elements/questionMark/questionMark.html"></require>
+  <require from="../tokenLabel.html"></require>
   <div class="page animated-page stageContainer stage3 au-animate ">
     <div class="infoContainer">
       <div class="heading heading2 title">Token Details</div>
