+++ src/newLaunch/lbp/stage4.ts
@@ -2,7 +2,7 @@ import { WhiteListService } from "services/WhiteListService";
 import { autoinject } from "aurelia-framework";
 import { Router } from "aurelia-router";
 import { DateService } from "services/DateService";
-import { BaseStage } from "newLbp/baseStage";
+import { BaseStage } from "newLaunch/baseStage";
 import Litepicker from "litepicker";
 import { Utils } from "services/utils";
 import { EventAggregator } from "aurelia-event-aggregator";
