+++ src/newLaunch/lbp/stage6.ts
@@ -1,5 +1,5 @@
 import { autoinject } from "aurelia-framework";
-import { BaseStage } from "newLbp/baseStage";
+import { BaseStage } from "newLaunch/baseStage";
 
 @autoinject
 export class Stage6 extends BaseStage {
